MLOps/legacy/fastapi-streamlit-ia-app/main/requirements.txt
fastapi==0.115.14
matplotlib==3.10.3
numpy==2.1.3
pandas==2.3.0
pillow==11.3.0
tensorflow==2.19.0
tensorflow-metal==1.2.0
